# Vulkan: VK_KHR_cooperative_matrix support to speed up prompt processing