This PR implements basic tensor core/matrix core/xmx engine support using the vendor-neutral `VK_KHR_cooperative_matrix` Vulkan extension for the matrix multiplication shader. I need to spend some further time optimizing the code, but I would like to get some feedback already.

This is different from #10206 in that it should work on some non-Nvidia GPUs too.

It works on:
- Nvidia RTX 2000 and newer

It should work on:
- AMD RX 7000 series and newer
- Intel ARC

HOWEVER:
It seems AMD's and Intel's drivers like to pretend they support this extension, despite either not having implemented proper hardware support (like my Intel A770 on Linux mesa), or not even having the necessary hardware at all (AMD's proprietary driver for older generations.. why?). So this probably also needs a whitelist feature to avoid regressions on hardware where support isn't actually there.

- [x] I have read the [contributing guidelines](https://github.com/ggerganov/llama.cpp/blob/master/CONTRIBUTING.md)
- Self-reported review complexity:
  - [ ] Low
  - [ ] Medium
  - [x] High