@0cc4m 2024-11-30
Some preliminary benchmarks, but there's probably some more optimization that can be done.

## RTX 3090
| model                          |       size |     params | backend    | ngl |          test |                  t/s master | t/s PR | t/s CUDA |
| ------------------------------ | ---------: | ---------: | ---------- | --: | ------------: | -------------------: | ------: | -------: |
| llama 7B Q4_0                  |   3.83 GiB |     7.24 B | Vulkan     |  99 |         pp512 |        825.42 ± 1.14 | 1933.18 ± 9.55 | 5059.05 ± 11.44 |
| llama 8B Q4_K - Small          |   4.36 GiB |     8.03 B | Vulkan     |  99 |         pp512 |        680.42 ± 0.73 | 1381.26 ± 7.01 | 4692.44 ± 4.78 |

Intel just for fun:

## Intel A770 (Linux Mesa)
| model                          |       size |     params | backend    | ngl |          test |                  t/s master | t/s PR |
| ------------------------------ | ---------: | ---------: | ---------- | --: | ------------: | -------------------: | ----------:|
| llama 8B Q4_K - Small          |   4.36 GiB |     8.03 B | Vulkan     |  99 |         pp512 |        150.46 ± 0.05 | 25.48 ± 0.12 |