@qnixsynapse 2024-11-30
> It seems AMD's and Intel's drivers like to pretend they support this extension, despite either not having implemented proper hardware support (like my Intel A770 on Linux mesa),

This should be reported to Freedesktop Gitlab IMO.

Also, I do test the Vulkan backend on my Intel Arc, I found out that Vulkan backend only uses the shader engines instead of compute engines. The later being faster. Is this intentional?