@0cc4m 2024-11-30
> > It seems AMD's and Intel's drivers like to pretend they support this extension, despite either not having implemented proper hardware support (like my Intel A770 on Linux mesa),
> 
> This should be reported to Freedesktop Gitlab IMO.

[They know.](https://gitlab.freedesktop.org/mesa/mesa/-/issues/9250)

> Also, I do test the Vulkan backend on my Intel Arc, I found out that Vulkan backend only uses the shader engines instead of compute engines. The later being faster. Is this intentional?

Assuming you mean shader cores and XMX engines, that's exactly what this PR is about. By default Vulkan can run compute shaders only on regular shader cores, it needs extensions like `VK_KHR_cooperative_matrix` to use the specialized hardware of tensor cores, AMD matrix cores or Intel XMX engines.