@jeffbolznv 2024-11-30
Very cool to see this! I tested the changes on RTX 4070, and test-backend-ops is passing, but I don't see much perf gain as-is (like +/- 20% testing a Q4_K model).

I see a couple perf opportunities. First, the branches around the loads for quants like Q4_K can cause the shader to stall inside the if/else waiting for the data. I had to "flatten" the control flow (e.g. https://github.com/ggerganov/llama.cpp/pull/10206/files#diff-508f8330faff804d4067ccb566c1d3eb27a7021a03df4f914c00c977d114514fR181) for the coopmat2 dequant functions, I suspect something similar is needed here.

Second, add `[[unroll]]` around the store loops. The dynamic indexing of the `sums` array is problematic and goes away with unrolling.

I tried these locally/hackily and then I get a much better speedup (like 2.5x).