@0cc4m 2024-11-30
> I see a couple perf opportunities. First, the branches around the loads for quants like Q4_K can cause the shader to stall inside the if/else waiting for the data. I had to "flatten" the control flow (e.g. https://github.com/ggerganov/llama.cpp/pull/10206/files#diff-508f8330faff804d4067ccb566c1d3eb27a7021a03df4f914c00c977d114514fR181) for the coopmat2 dequant functions, I suspect something similar is needed here.

Interesting. So that the GPU has to wait for the load from global memory, but this is worsened by being in a branch?

> Second, add `[[unroll]]` around the store loops. The dynamic indexing of the `sums` array is problematic and goes away with unrolling.
> 
> I tried these locally/hackily and then I get a much better speedup (like 2.5x).

I implemented those changes. I hope you don't mind that I used your loading code. Maybe in the future we can just use your dequant_funcs file for mul_mm too.

I don't see a drastic improvement, but it's 7-8% faster for me on 3090 on Q4_K_S, so that's good. Let me know if you see anything else.

I'm also curious if you learned these GLSL optimization techniques from experience, or if you can recommend some resources I can use to learn. This is my first Vulkan project and the first time going deep into GPU programming, and it's been difficult to find information I can apply.