@jeffbolznv 2024-11-30
> Interesting. So that the GPU has to wait for the load from global memory, but this is worsened by being in a branch?

Yeah, it's challenging for a compiler to track memory accesses across basic blocks, so doing loads inside a branch can cause the value to be waited on inside the branch. And that prevents doing other useful work while the load is pending.

> I'm also curious if you learned these GLSL optimization techniques from experience, or if you can recommend some resources I can use to learn. This is my first Vulkan project and the first time going deep into GPU programming, and it's been difficult to find information I can apply.

I think most performance advice is likely to come from vendor specific documentation/tooling, even though a lot of the information and optimization techniques can apply across different hardware. I'd suggest looking at nsight documentation and cuda forums as they'll likely have a lot more low-level information about NVIDIA hardware than you would find in Vulkan-centric documentation. For example, https://developer.nvidia.com/blog/identifying-shader-limiters-with-the-shader-profiler-in-nvidia-nsight-graphics/ talks about using nsight and also touches on this issue of dynamically indexed arrays.