@jeffbolznv 2024-11-30
I got a device lost error while trying to collect perf results, running Meta-Llama-3-8B-Instruct-Q4_K_S.gguf with pp512. But these two other models run:

```
master
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | Vulkan     | 1000 |         pp512 |        831.12  4.03 |
| phi3 3B Q4_K - Medium          |   2.23 GiB |     3.82 B | Vulkan     | 1000 |         pp512 |       1135.54  1.70 |

PR
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | Vulkan     | 1000 |         pp512 |       1904.89  8.26 |
| phi3 3B Q4_K - Medium          |   2.23 GiB |     3.82 B | Vulkan     | 1000 |         pp512 |      2213.51  25.68 |
```