@characharm 2024-12-01
Windows OS

this pr
ggml_vulkan: Found 1 Vulkan devices:
ggml_vulkan: 0 = Intel(R) Arc(TM) A770 Graphics (Intel Corporation) | uma: 0 | fp16: 1 | warp size: 32 | matrix cores: 1
| model                          |       size |     params | backend    | ngl |          test |                  t/s |
| ------------------------------ | ---------: | ---------: | ---------- | --: | ------------: | -------------------: |
ggml_vulkan: Compiling shaders..............................Done!
| qwen2 14B Q5_K - Medium        |   9.78 GiB |    14.77 B | Vulkan     |  99 |         pp512 |         20.84 ± 0.03 |
| qwen2 14B Q5_K - Medium        |   9.78 GiB |    14.77 B | Vulkan     |  99 |         tg128 |         16.01 ± 0.06 |

build: 56609765 (4231)

master 
ggml_vulkan: Found 1 Vulkan devices:
ggml_vulkan: 0 = Intel(R) Arc(TM) A770 Graphics (Intel Corporation) | uma: 0 | fp16: 1 | warp size: 32
| model                          |       size |     params | backend    | ngl |          test |                  t/s |
| ------------------------------ | ---------: | ---------: | ---------- | --: | ------------: | -------------------: |
ggml_vulkan: Compiling shaders..............................Done!
| qwen2 ?B Q5_K - Medium         |   9.78 GiB |    14.77 B | Vulkan,RPC |  99 |         pp512 |         67.85 ± 0.52 |
| qwen2 ?B Q5_K - Medium         |   9.78 GiB |    14.77 B | Vulkan,RPC |  99 |         tg128 |         15.52 ± 0.06 |

build: 25669aa9 (4179)